New migration for teacherInfo and some minor fixes and moving files
